OCR/requirements.txt
google-cloud-vision
imutils
numpy
opencv-contrib-python
Pillow